Task 02(vectorMirrorImage) test

+++ src/by/itstep/java1022/belser/model/logic/VectorSort.java
@@ -1,7 +1,7 @@
 package by.itstep.java1022.belser.model.logic;
 
 public class VectorSort {
-    public static String sort(int[] vector) {
+    public static String sort(double[] vector) {
         String msg = "Vector is sorted";
         for (int i = 1; i < vector.length - 1; i++) {
             if ((vector[i - 1] < vector[i] && vector[i] > vector[i + 1]) ||
